.claude-plugin/skills/daily-briefing/requirements.txt
# 每日简报技能依赖

# RSS源抓取
feedparser>=6.0.10

# 网页抓取
requests>=2.31.0
beautifulsoup4>=4.12.0

# 配置文件解析
PyYAML>=6.0.0
